requirements.txt
chromadb
langchain
langchain-community
langchain-experimental
langchain-openai
langgraph
langsmith
matplotlib
langchain-anthropic
numexpr
numpy
openai
pandas
Pillow
python-dotenv
requests
sql-metadata
sqlalchemy
torch
transformers
tqdm
typing_extensions
